fix(acp): redact launch args from runtime log and honor file-size ratchet
Review item #1 (security): spawn_agent_child logged the raw agent_args
vector with `{:?}` into the runtime log retrievable end-to-end via
get_managed_agent_log. Args may legally carry credentials
(--token=...), so no argument value may ever be serialized. The
resolved-launch marker now records only the resolved command and the
argument count, mirroring spawn_snapshot::diff's MaskedBare policy.

The marker writer lives in storage.rs as
append_resolved_launch_marker(path, command, args): it resolves the
command to a full path (DMG launches have a minimal PATH), appends the
redacted marker, and returns the resolved command for the
BUZZ_ACP_AGENT_COMMAND spawn env, replacing runtime.rs's inline
resolution. open_log_file now creates runtime logs with mode 0o600 at
creation time (Unix), so the permissions hold regardless of umask;
agent stdout/stderr can echo installer/CLI credentials.

Regression tests: append_resolved_launch_marker_never_writes_argument_
values asserts the marker carries the command and args_count only --
never token/argument values -- and that unresolvable commands fall
through verbatim; open_log_file_creates_owner_only asserts 0o600.

File-size ratchet: the two effective-launch tests moved from
readiness.rs into readiness_effective_launch_tests.rs (external
#[path] module, following readiness_goose_file_config_tests.rs), and
runtime.rs stays at/below the ratchet cap after folding command
resolution into the storage helper. Unblocks Detect Changed Paths so
the required CI jobs run.
